refactor: reduce duplication
